+++ scm/domain-service/main/src/api/mod.rs
@@ -7,8 +7,8 @@ pub use service::ServiceRegistry;
 pub use service::ServiceError;
 pub use service::{
     EmptinessRequest, EmptinessResponse, LenRequest, LenResponse, ListNamesRequest,
-    ListNamesResponse, NameRequest, NameResponse, NoopRequest, NoopResponse, NoopService,
-    RegisterServiceRequest, RegisterServiceResponse, Request, Response, ServiceLookupRequest,
-    ServiceLookupResponse, ServiceRegistryStore, ServiceRemovalRequest, ServiceRemovalResponse,
+    ListNamesResponse, NameRequest, NameResponse, NoopService, RegisterServiceRequest,
+    RegisterServiceResponse, Request, Response, ServiceLookupRequest, ServiceLookupResponse,
+    ServiceRegistryStore, ServiceRemovalRequest, ServiceRemovalResponse,
     StdServiceRegistryFactory,
 };