+++ scm/domain-service/main/src/api/tests/service/service_registry_e2e_test.rs
@@ -3,21 +3,21 @@
 #[cfg(test)]
 mod tests {
     use crate::api::{
-        EmptinessRequest, LenRequest, ListNamesRequest, NoopRequest, NoopResponse, NoopService,
-        RegisterServiceRequest, Service, ServiceError, ServiceLookupRequest, ServiceRegistry,
-        ServiceRemovalRequest,
+        EmptinessRequest, LenRequest, ListNamesRequest, NoopService, RegisterServiceRequest,
+        Service, ServiceError, ServiceLookupRequest, ServiceRegistry, ServiceRemovalRequest,
     };
+    use edge_application_base::{EmptyRequest, EmptyResponse};
     use std::sync::Arc;
 
     struct TestRegistry;
 
     impl ServiceRegistry for TestRegistry {
-        type Request = NoopRequest;
-        type Response = NoopResponse;
+        type Request = EmptyRequest;
+        type Response = EmptyResponse;
 
         fn register(
             &self,
-            _req: &RegisterServiceRequest<NoopRequest, NoopResponse>,
+            _req: &RegisterServiceRequest<EmptyRequest, EmptyResponse>,
         ) -> Result<crate::api::RegisterServiceResponse, ServiceError> {
             Ok(crate::api::RegisterServiceResponse)
         }
@@ -32,7 +32,7 @@ mod tests {
         fn get(
             &self,
             _req: &ServiceLookupRequest,
-        ) -> Result<crate::api::ServiceLookupResponse<NoopRequest, NoopResponse>, ServiceError> {
+        ) -> Result<crate::api::ServiceLookupResponse<EmptyRequest, EmptyResponse>, ServiceError> {
             Ok(crate::api::ServiceLookupResponse { service: None })
         }
 
@@ -62,7 +62,7 @@ mod tests {
             crate::api::NoopService
         }
 
-        fn new_store() -> crate::api::ServiceRegistryStore<NoopRequest, NoopResponse> {
+        fn new_store() -> crate::api::ServiceRegistryStore<EmptyRequest, EmptyResponse> {
             crate::api::ServiceRegistryStore::default()
         }
     }